+++ app/controllers/analytics/subject_controller.rb
@@ -4,54 +4,12 @@ require 'sql/queries'
 include SQL # rubocop:disable Style/MixinUsage
 
 module Analytics
-  # rubocop:disable Metrics/ClassLength
   class SubjectController < AnalyticsController
     # rubocop:disable Metrics/PerceivedComplexity
     # rubocop:disable Metrics/MethodLength
     # rubocop:disable Metrics/CyclomaticComplexity
     # rubocop:disable Metrics/AbcSize
     def index
-      @subject = params[:subject_select]
-
-      @organizations = policy_scope Organization
-      @chapters = if !@selected_organization_id.nil? && (@selected_organization_id != '') && (@selected_organization_id != 'All')
-                    policy_scope Chapter.where(organization_id: @selected_organization_id)
-                  else
-                    policy_scope Chapter
-                  end
-
-      @groups = if !@selected_group_id.nil? && (@selected_group_id != '') && (@selected_group_id != 'All')
-                  policy_scope Group.where(chapter_id: @selected_chapter_id)
-                elsif !@selected_organization_id.nil? && (@selected_organization_id != '') && (@selected_organization_id != 'All')
-                  policy_scope Group.includes(:chapter).where(chapters: { organization_id: @selected_organization_id })
-                else
-                  policy_scope Group
-                end
-
-      @students = if !@selected_group_id.nil? && (@selected_group_id != '') && (@selected_group_id != 'All')
-                    policy_scope Student.where(group_id: @selected_group_id).order(:last_name, :first_name)
-                  elsif !@selected_chapter_id.nil? && (@selected_chapter_id != '') && (@selected_chapter_id != 'All')
-                    policy_scope Student.includes(:group).where(groups: { chapter_id: @selected_chapter_id }).order(:last_name, :first_name)
-                  elsif !@selected_organization_id.nil? && (@selected_organization_id != '') && (@selected_organization_id != 'All')
-                    policy_scope Student.includes(group: :chapter).where(chapters: { organization_id: @selected_organization_id }).order(:last_name, :first_name)
-                  else
-                    policy_scope Student.order(:last_name, :first_name)
-                  end
-
-      @subjects = if !@selected_organization_id.nil? && (@selected_organization_id != '') && (@selected_organization_id != 'All')
-                    policy_scope Subject.where(organization: @selected_organization_id)
-                  else
-                    policy_scope Subject
-                  end
-
-      @selected_organization_id = @organizations.first.id unless params[:organization_select]
-      @subject = @subjects.first.id unless params[:subject_select]
-
-      # figure 3: Histograms (Trellis) for the seven skills that are evaluated
-      # multiple series (1 per group) per histogram;
-      # x-axis: nr. of lessons
-      # y-axis: average score
-      # series = [{skill : skill_name, series : [{name : group_name, data : [[x, y], ..]}]}]
       series3 = if @selected_student_id.present? && @selected_student_id != 'All'
                   performance_per_skill_single_student
                 else
@@ -151,5 +109,4 @@ module Analytics
     # rubocop:enable Metrics/PerceivedComplexity
     # rubocop:enable Metrics/CyclomaticComplexity
   end
-  # rubocop:enable Metrics/ClassLength
 end